+++ src/routes/catalog/catalog.test.js
@@ -1,0 +1,6 @@
+import app from '../../app';
+import { testRequest } from '../../helpers';
+
+describe('Catalog', () => {
+  it('should return the catalog info', () => testRequest(app, 'catalog/get'));
+});